przypisanie tasków do użytkownika

+++ controllers/tasks.php
@@ -1,11 +1,15 @@
 <?php
 
 function getTask(){
-    return dbQuery("SELECT * FROM task_list ORDER BY Date ASC");
+    return dbQuery("SELECT * FROM task_list WHERE id = '" . getLoginId() . "' ORDER BY Date ASC");
 }
 
 function addTask($title,$status,$group){
    
+   $userID = getLoginId();
+   if ($userID === -1) {
+       return ['Status' => 'Error', 'Description' => 'Nie jesteś zalogowany!'];
+   }
    if ($status > 3 && $status < 0) {
        return ['Status' => 'Error', 'Description' => 'Status może mieć wartość z zakresu 0-3, przekazano :'. $status];
    }
@@ -16,14 +20,16 @@ function addTask($title,$status,$group){
    if (!isset($allowedGroup[$group])) {
        return ['Status' => 'Error', 'Description' => 'Wybrano grupę spoza zakresu '];
     }
-     dbQuery("INSERT INTO `task_list` (`ID`, `Title`, `Description`, `Date`, `Status`, `Groups`) VALUES (NULL, '$title', '', NOW(), '$status', '$group');");
+     dbQuery("INSERT INTO `task_list` (`ID`,`User_id`, `Title`, `Description`, `Date`, `Status`, `Groups`) VALUES (NULL, '$userID', '$title', '', NOW(), '$status', '$group');");
      return ['Status' => 'OK', 'Description' => 'Wszystko OK'];
 }
 
 function removeTask($id){
-    return dbQuery("DELETE FROM task_list WHERE id = $id");
+    $userID = getLoginId();
+    return dbQuery("DELETE FROM task_list WHERE id = '$id' AND User_id = '$userID'");
 }
 
 function editTask($id, $title, $status, $groups){
-    return dbQuery("UPDATE task_list SET Title='$title', Status='$status', Groups='$groups' WHERE id=$id");
+    $userID = getLoginId();
+    return dbQuery("UPDATE task_list SET Title='$title', Status='$status', Groups='$groups' WHERE id = '$id' AND User_id = '$userID'");
 }
